+++ jdaviz/components/plugin_add_results.vue
@@ -54,11 +54,21 @@
 
     <v-row justify="end">
       <j-tooltip :tooltipcontent="label_overwrite ? action_tooltip+' and replace existing entry' : action_tooltip">
-        <v-btn :disabled="label_invalid_msg.length > 0 || action_disabled"
+        <v-btn :disabled="label_invalid_msg.length > 0 || action_disabled || action_spinner"
           text
           color="accent"
           @click="$emit('click:action')"
-        >{{action_label}}{{label_overwrite ? ' (Overwrite)' : ''}}
+        >
+          <v-progress-circular
+            v-if="action_spinner"
+            indeterminate
+            color="primary"
+            size="20"
+            width="3"
+            style="margin-right: 4px"
+          ></v-progress-circular>
+          <span v-else style="width: 24px"></span>
+          {{action_label}}{{label_overwrite ? ' (Overwrite)' : ''}}
         </v-btn>
       </j-tooltip>
     </v-row>
@@ -75,6 +85,6 @@
 module.exports = {
   props: ['label', 'label_default', 'label_auto', 'label_invalid_msg', 'label_overwrite', 'label_label', 'label_hint',
           'add_to_viewer_items', 'add_to_viewer_selected', 'add_to_viewer_hint',
-          'action_disabled', 'action_label', 'action_tooltip']
+          'action_disabled', 'action_spinner', 'action_label', 'action_tooltip']
 };
 </script>